Organized components to logical named folders
