+++ mobile/src/components/home/VendorCard.jsx
