+++ mobile/src/components/scan/CustomBottomSheet.jsx
