+++ mobile/src/components/scan/CustomScanIcon.jsx
